requirements.txt
flask>=3.0.0
flask-wtf>=1.2.0
PyYAML>=6.0.1
requests>=2.31.0
certifi>=2024.0.0
